BFF - fix occasionally failing dpop test
