+++ bff/test/Duende.Bff.Tests/Endpoints/RemoteEndpointTests.cs
@@ -3,15 +3,11 @@
 
 using Duende.Bff.Tests.TestFramework;
 using Duende.Bff.Tests.TestHosts;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.IdentityModel.Tokens;
 using System;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Json;
-using System.Security.Cryptography;
-using System.Text.Json;
 using System.Threading.Tasks;
 using Shouldly;
 using Xunit;
@@ -314,29 +310,5 @@ namespace Duende.Bff.Tests.Endpoints
             await f.ShouldThrowAsync<Exception>();
         }
 
-        [Fact]
-        public async Task test_dpop()
-        {
-            var rsaKey = new RsaSecurityKey(RSA.Create(2048));
-            var jsonWebKey = JsonWebKeyConverter.ConvertFromRSASecurityKey(rsaKey);
-            jsonWebKey.Alg = "PS256";
-            var jwk = JsonSerializer.Serialize(jsonWebKey);
-
-            BffHost.OnConfigureServices += svcs =>
-            {
-                svcs.PostConfigure<BffOptions>(opts =>
-                {
-                    opts.DPoPJsonWebKey = jwk;
-                });
-            };
-            await BffHost.InitializeAsync();
-
-            var apiResult = await BffHost.BrowserClient.CallBffHostApi(
-                url: BffHost.Url("/api_client/test")
-            );
-
-            apiResult.RequestHeaders["DPoP"].First().ShouldNotBeNullOrEmpty();
-            apiResult.RequestHeaders["Authorization"].First().StartsWith("DPoP ").ShouldBeTrue();
-        }
     }
 }
